+++ src/kelary/IviVisaCompatibiityTester/Program.cs
@@ -15,8 +15,8 @@ if ( args is not null && args.Length > 0 )
     Console.WriteLine();
     resourceName = args[0];
 
-    Console.WriteLine();
-    Console.WriteLine( $"Make sure that the instrument at {resourceName} is turned on." );
+    Console.Write( $"\nTurn on the instrument at {resourceName} and press any key »" );
+    _ = Console.ReadKey();
 }
 
 TargetFrameworkAttribute? framework = Assembly.GetEntryAssembly()?.GetCustomAttribute<TargetFrameworkAttribute>();
@@ -72,6 +72,5 @@ if ( !string.IsNullOrWhiteSpace( resourceName ) )
         Console.WriteLine( details );
 }
 
-Console.WriteLine();
-Console.WriteLine( "Press any key to finish." );
+Console.Write( "\nPress any key to finish »" );
 _ = Console.ReadKey();